chore(gfql): disambiguate reentry compile-time module naming
